+++ braid/braid_geweke/src/lib.rs
@@ -1,4 +1,5 @@
 //! Geweke (joint distribution) test
+#![warn(unused_extern_crates)]
 #![warn(
     clippy::all,
     clippy::imprecise_flops,