+++ braid/src/config.rs
@@ -37,10 +37,29 @@ impl EngineUpdateConfig {
         Self {
             n_iters: 1,
             timeout: None,
-            transitions: DEFAULT_STATE_TRANSITIONS.into(),
+            transitions: Vec::new(),
             save_config: None,
             checkpoint: None,
         }
+    }
+
+    pub fn with_default_transitions() -> Self {
+        Self::new().default_transitions()
+    }
+
+    pub fn default_transitions(mut self) -> Self {
+        self.transitions = DEFAULT_STATE_TRANSITIONS.into();
+        self
+    }
+
+    pub fn transitions(mut self, transitions: Vec<StateTransition>) -> Self {
+        self.transitions.extend(transitions);
+        self
+    }
+
+    pub fn transition(mut self, transition: StateTransition) -> Self {
+        self.transitions.push(transition);
+        self
     }
 
     /// Emit a `StateUpdateConfig` with the same settings
@@ -50,6 +69,21 @@ impl EngineUpdateConfig {
             timeout: self.timeout,
             transitions: self.transitions.clone(),
         }
+    }
+
+    pub fn n_iters(mut self, n_iters: usize) -> Self {
+        self.n_iters = n_iters;
+        self
+    }
+
+    pub fn checkpoint(mut self, checkpoint: Option<usize>) -> Self {
+        self.checkpoint = checkpoint;
+        self
+    }
+
+    pub fn timeout(mut self, seconds: Option<u64>) -> Self {
+        self.timeout = seconds;
+        self
     }
 }