+++ braid/tests/engine.rs
@@ -136,7 +136,9 @@ fn update_empty_engine_smoke_test() {
     )
     .unwrap();
 
-    engine.update(EngineUpdateConfig::default(), None).unwrap();
+    engine
+        .update(EngineUpdateConfig::with_default_transitions(), None, None)
+        .unwrap();
 }
 
 #[test]
@@ -295,17 +297,17 @@ mod prior_in_codebook {
         ---
         table_name: table
         state_alpha_prior:
-            Gamma:
+            !Gamma
                 shape: 1.0
                 rate: 1.0
         view_alpha_prior:
-            Gamma:
+            !Gamma
                 shape: 1.0
                 rate: 1.0
         col_metadata:
             - name: x
               coltype:
-                Continuous:
+                !Continuous
                     prior:
                         m: 0.0
                         k: 1.0
@@ -313,7 +315,7 @@ mod prior_in_codebook {
                         s2: 3.0
             - name: y
               coltype:
-                Continuous:
+                !Continuous
                     hyper: ~
                     prior: ~
         comments: ~
@@ -1477,25 +1479,25 @@ mod insert_data {
         assert_eq!(engine.n_rows(), 3);
         assert_eq!(engine.n_cols(), 1);
 
-        engine.update(cfg.clone(), None).unwrap();
+        engine.update(cfg.clone(), None, None).unwrap();
 
         add_row(&mut engine, "b1", 1.0).unwrap();
 
         assert_eq!(engine.n_rows(), 4);
         assert_eq!(engine.n_cols(), 1);
-        engine.update(cfg.clone(), None).unwrap();
+        engine.update(cfg.clone(), None, None).unwrap();
         assert_eq!(engine.n_rows(), 4);
 
         add_row(&mut engine, "b2", -1.0).unwrap();
 
         assert_eq!(engine.n_rows(), 5);
-        engine.update(cfg.clone(), None).unwrap();
+        engine.update(cfg.clone(), None, None).unwrap();
         assert_eq!(engine.n_rows(), 5);
 
         add_row(&mut engine, "b3", 0.0).unwrap();
 
         assert_eq!(engine.n_rows(), 6);
-        engine.update(cfg, None).unwrap();
+        engine.update(cfg, None, None).unwrap();
         assert_eq!(engine.n_rows(), 6);
     }
 
@@ -1573,25 +1575,25 @@ mod insert_data {
         assert_eq!(engine.n_rows(), 3);
         assert_eq!(engine.n_cols(), 2);
 
-        engine.update(cfg.clone(), None).unwrap();
+        engine.update(cfg.clone(), None, None).unwrap();
 
         add_row(&mut engine, "b1", 1.0, 0.5).unwrap();
 
         assert_eq!(engine.n_rows(), 4);
         assert_eq!(engine.n_cols(), 2);
-        engine.update(cfg.clone(), None).unwrap();
+        engine.update(cfg.clone(), None, None).unwrap();
         assert_eq!(engine.n_rows(), 4);
 
         add_row(&mut engine, "b2", -1.0, 0.1).unwrap();
 
         assert_eq!(engine.n_rows(), 5);
-        engine.update(cfg.clone(), None).unwrap();
+        engine.update(cfg.clone(), None, None).unwrap();
         assert_eq!(engine.n_rows(), 5);
 
         add_row(&mut engine, "b3", 0.0, -1.2).unwrap();
 
         assert_eq!(engine.n_rows(), 6);
-        engine.update(cfg, None).unwrap();
+        engine.update(cfg, None, None).unwrap();
         assert_eq!(engine.n_rows(), 6);
     }
 
@@ -1907,6 +1909,7 @@ mod insert_data {
                             ],
                             ..Default::default()
                         },
+                        None,
                         None,
                     )
                     .unwrap();
@@ -2315,7 +2318,7 @@ mod insert_data {
                     ..Default::default()
                 };
 
-                engine.update(cfg, None).unwrap();
+                engine.update(cfg, None, None).unwrap();
 
                 assert_eq!(engine.n_rows(), starting_rows);
             }